Re-organize sources
