+++ include/sir/maps.h
@@ -1,5 +1,5 @@
 /*
- * sirconsole.h
+ * maps.h
  *
  * Author:    Ryan M. Lederman <lederman@gmail.com>
  * Copyright: Copyright (c) 2018-2023
@@ -23,42 +23,12 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-#ifndef _SIR_CONSOLE_H_INCLUDED
-# define _SIR_CONSOLE_H_INCLUDED
+#ifndef _SIR_MAPS_H_INCLUDED
+# define _SIR_MAPS_H_INCLUDED
 
-# include "sirtypes.h"
-# include "sirhelpers.h"
+# include "sir/types.h"
 
-# if !defined(__WIN__)
-bool _sir_write_stdio(FILE* stream, const char* message);
+extern sir_level_style_tuple sir_level_to_style_map[SIR_NUMLEVELS];
+extern sir_level_str_pair sir_level_to_str_map[SIR_NUMLEVELS];
 
-static inline
-bool _sir_write_stdout(const char* message, size_t len) {
-    _SIR_UNUSED(len);
-    return _sir_write_stdio(stdout, message);
-}
-
-static inline
-bool _sir_write_stderr(const char* message, size_t len) {
-    _SIR_UNUSED(len);
-    return _sir_write_stdio(stderr, message);
-}
-# else /* __WIN__ */
-extern HANDLE __sir_stdout;
-extern HANDLE __sir_stderr;
-
-bool _sir_initialize_stdio(void);
-bool _sir_write_stdio(HANDLE console, const char* message, size_t len);
-
-static inline
-bool _sir_write_stdout(const char* message, size_t len) {
-    return _sir_write_stdio(__sir_stdout, message, len);
-}
-
-static inline
-bool _sir_write_stderr(const char* message, size_t len) {
-    return _sir_write_stdio(__sir_stderr, message, len);
-}
-# endif // !__WIN__
-
-#endif /* !_SIR_CONSOLE_H_INCLUDED */
+#endif // !_SIR_MAPS_H_INCLUDED
