+++ src/sirconsole.c
@@ -23,8 +23,8 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-#include "sirconsole.h"
-#include "sirinternal.h"
+#include "sir/console.h"
+#include "sir/internal.h"
 
 #if !defined(__WIN__)
