+++ src/sirfilesystem.c
@@ -23,8 +23,8 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-#include "sirfilesystem.h"
-#include "sirinternal.h"
+#include "sir/filesystem.h"
+#include "sir/internal.h"
 
 #if defined(__WIN__)
 # pragma comment(lib, "Shlwapi.lib")
