+++ src/sirhelpers.c
@@ -23,8 +23,8 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-#include "sirhelpers.h"
-#include "sirerrors.h"
+#include "sir/helpers.h"
+#include "sir/errors.h"
 
 void __sir_safefree(void** pp) {
     if (!pp || !*pp)
